refactor: Combination rules refdata find
Added find refdata value function to combination rules resource
