+++ test/resources/utils/index.js
@@ -1,0 +1,1 @@
+export { default as findRefdataValue } from './findRefdataValue';
